pyproject.toml
[project]
name = "mlops_cml"
version = "0.1.0"
description = "Entregable asignatura aprendizaje automatico en la nube"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gunicorn>=23.0.0",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.5",
    "mlflow>=3.2.0",
    "numpy>=2.3.2",
    "optuna>=4.4.0",
    "pandas>=2.3.1",
    "prefect>=3.4.12",
    "pyarrow>=21.0.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "xgboost>=3.0.4",
    "kagglehub",
    "holidays",
    "statsmodels",
    "files",
    "storage",
    "google-cloud-storage",
    "optuna-integration[mlflow]",
    "tensorflow",
    "s3fs",
]

[tool.uv.workspace]
members = [
    "04-Deployment/deploy/web-service",
    "04-Deployment/deploy/web-service/taxi-prediction",
]
